@@ -23,11 +23,11 @@
   "author": "Dynamic <dev@dynamicagency.com>",
   "license": "BSD-3-Clause",
   "dependencies": {
-    "@fortawesome/fontawesome": "^1.1.4",
-    "@fortawesome/fontawesome-free-solid": "^5.0.8",
-    "@fortawesome/react-fontawesome": "0.0.17",
-    "animated-scroll-to": "^1.1.8",
-    "axios": "^0.17.1",
+    "@fortawesome/fontawesome": "^1.1.5",
+    "@fortawesome/fontawesome-free-solid": "^5.0.9",
+    "@fortawesome/react-fontawesome": "0.0.18",
+    "animated-scroll-to": "^1.1.10",
+    "axios": "^0.18.0",
     "bootstrap": "^4.0.0",
     "es6-promise": "^4.2.4",
     "handlebars": "^4.0.11",
@@ -40,7 +40,7 @@
     "react-document-title": "^2.0.3",
     "react-dom": "^16.2.0",
     "react-google-maps": "^9.4.5",
-    "react-places-autocomplete": "^6.1.2",
+    "react-places-autocomplete": "^6.1.3",
     "react-redux": "^5.0.7",
     "react-virtualized": "^9.18.5",
     "redux": "^3.7.2",
@@ -48,20 +48,20 @@
     "redux-thunk": "^2.2.0"
   },
   "devDependencies": {
-    "autoprefixer": "^7.2.6",
+    "autoprefixer": "^8.2.0",
     "axios-mock-adapter": "^1.14.1",
     "babel-core": "^6.26.0",
-    "babel-jest": "^22.4.1",
-    "babel-loader": "^7.1.3",
+    "babel-jest": "^22.4.3",
+    "babel-loader": "^7.1.4",
     "babel-plugin-transform-object-rest-spread": "^6.26.0",
     "babel-preset-env": "^1.6.1",
     "babel-preset-react": "^6.11.1",
     "babel-preset-stage-2": "^6.24.1",
     "babel-register": "^6.26.0",
-    "css-loader": "^0.28.10",
+    "css-loader": "^0.28.11",
     "enzyme": "^3.3.0",
     "enzyme-adapter-react-16": "^1.1.1",
-    "eslint": "^4.18.2",
+    "eslint": "^4.19.1",
     "eslint-config-airbnb": "^16.1.0",
     "eslint-config-airbnb-base": "^12.1.0",
     "eslint-import-resolver-node": "^0.3.2",
@@ -70,16 +70,16 @@
     "eslint-plugin-react": "^7.7.0",
     "extract-text-webpack-plugin": "^3.0.2",
     "file-loader": "^1.1.11",
-    "jest": "^22.4.2",
-    "jest-cli": "^22.4.2",
-    "jest-enzyme": "^4.2.0",
-    "node-sass": "^4.7.2",
-    "postcss-loader": "^2.1.1",
+    "jest": "^22.4.3",
+    "jest-cli": "^22.4.3",
+    "jest-enzyme": "^6.0.0",
+    "node-sass": "^4.8.3",
+    "postcss-loader": "^2.1.3",
     "react-test-renderer": "^16.2.0",
     "remote-redux-devtools": "^0.5.12",
-    "resolve-url-loader": "^2.2.1",
-    "sass-loader": "^6.0.5",
-    "style-loader": "^0.20.2",
+    "resolve-url-loader": "^2.3.0",
+    "sass-loader": "^6.0.7",
+    "style-loader": "^0.20.3",
     "webpack": "^3.11.0"
   },
   "browserslist": "> 1%",
